requirements.txt
fastapi
uvicorn
torch
torchvision
timm
pillow
